bybit_futures_bot/requirements.txt
pybit==5.11.0
pandas==2.2.3
ta==0.11.0
python-dotenv==1.0.0
requests==2.31.0

